make character speed the same at any frame rate

+++ limit.cpp
@@ -1,8 +1,11 @@
-int limit (int value, int min, int max) {
-    if (value < min) {
+float limit(float value, float min, int max)
+{
+    if (value < min)
+    {
         return min;
     }
-    if (value > max) {
+    if (value > max)
+    {
         return max;
     }
     return value;
